Feature: full Anthropic Messages API for Claude Code + MCP compare tool
Anthropic Messages API (/v1/messages):
- Add tools, tool_choice, metadata, thinking fields to MessagesRequest
- Add tool_use, tool_result, thinking, redacted_thinking content blocks
- Full pass-through to Anthropic cloud (tools, thinking, cache_control preserved)
- Non-Claude models translated Anthropic→OpenAI and routed to cloud providers
- Tool results/calls/thinking converted to text for local inference
- ResponseContentBlock now enum with Text, ToolUse, Thinking variants
- SystemBlock.cache_control preserved for Anthropic prompt caching
- 6 new unit tests (tool_use, tool_result, thinking, tools request, serialization)

MCP compare tool:
- New 'compare' tool sends same prompt to multiple models concurrently
- Supports local, network, and cloud models (up to 10)
- Returns side-by-side results with latency_ms, token counts, status
- Self-requests through /v1/messages for consistent routing
- Claude Code users can compare model outputs via MCP

Claude Code integration:
  ANTHROPIC_BASE_URL=http://localhost:8800
  ANTHROPIC_AUTH_TOKEN=$(cat api_key)
  claude --model qwen2.5-coder-7b

MCP config for comparison mode:
  {"mcpServers":{"swarmllm":{"url":"http://localhost:8800/mcp"}}}

Also fixes: clippy unnecessary_cast in daemon.rs

+++ src/daemon.rs
@@ -2760,7 +2760,7 @@ fn regenerate_manifest_from_header(
         && !model_dir.join("shard_001.bin").exists()
         && shard0_path
             .metadata()
-            .map(|m| m.len() as u64 >= (total_size as u64).saturating_sub(4096))
+            .map(|m| m.len() >= total_size.saturating_sub(4096))
             .unwrap_or(false);
 
     let (shard_count, shards) = if is_single_full_gguf {
